+++ 1lvl/utils.h
@@ -1,8 +1,8 @@
 #ifndef __NB_ALLOC_UTILS__
 #define __NB_ALLOC_UTILS__
 
-unsigned rand_lim(unsigned limit);
-unsigned long upper_power_of_two(unsigned long v);
-unsigned log2_ (unsigned long value);
+unsigned int rand_lim(unsigned int limit);
+unsigned int upper_power_of_two(unsigned int v);
+unsigned int log2_ (unsigned int value);
 
 #endif
